more fixes for get articles content
