+++ Backend/src/services/remotecontentmanagerifs.ts
@@ -1,0 +1,3 @@
+export interface RemoteContentManagerIfs { 
+    getArcicleContent(articleName: string, language: string): Promise<string>;
+}
